+++ duva-client/src/cli/main.rs
@@ -1,8 +1,5 @@
 mod cli;
-pub mod completion;
 mod editor;
-mod hints;
-
 use clap::Parser;
 use duva::prelude::tokio::{self, sync::oneshot};
 use duva_client::{
